Support importance specifier on cell cards
This feature works the same way as material specification.
Cell cards that have imp:* (where * is a single character)
will generate cells that are grouped into CUBIT groups
with names like `imp.n_0.5`.  The importance types are not
restricted to the set (n,p,e), though those are the only
valid ones for MCNP.

imp cards in the data card section of the input are *not* parsed.

Unfortunately, some unrelated whitespace fixes got pulled into this
update as well.